% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BabyNames}
\alias{BabyNames}
\title{Names of children as recorded by the US Social Security Administration.}
\format{
A data frame with 1,792,091 entries on four variables:
\describe{
  \item{name}{The given name (character string)}
  \item{sex}{F or M (character string)}
  \item{count}{The number of babies given that name and of that sex. (integer)}
  \item{year}{Year of birth (integer)}
}
}
\source{
The data were compiled from the US Social Security Administration
web site: http://www.ssa.gov/oact/babynames/names.zip
}
\usage{
data("BabyNames")
}
\description{
The US Social Security Administration provides yearly lists of names given
to babies. These data combine the yearly lists.
}
\details{
`BabyNames` is the raw data from the SSA. The case is a year-name-sex, for
example: Jane F 1922. The count is the number of children of that sex given
that name in that year. Names assigned to fewer than five children of one
sex in any year are not listed, presumably out of privacy concerns.
}
\examples{
data(BabyNames)
str(BabyNames)
}
\keyword{datasets}
