% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountryCentroids}
\alias{CountryCentroids}
\title{Geographic locations of countries}
\format{
A data frame with 241 observations on the following 4 variables:
\describe{
  \item{name}{country name in English, e.g., Afghanistan, Zimbabwe}
  \item{iso_a3}{The ISO A3 code for the country, e.g. AFG, ZWE}
  \item{long}{Longitude in degrees}
  \item{lat}{Latitude in degrees}
}
}
\source{
The centroids were derived from country boundary map data as the
medians of the boundary points latitude and longitude.
}
\usage{
data("CountryCentroids")
}
\description{
Gives the latitude and longitude for many countries. A single point — the
centroid of the country's boundaries — is given. Countries are identified
by English name and by the ISO A3 code.
}
\examples{
data(CountryCentroids)
plot(lat ~ long, data=CountryCentroids)
}
\keyword{datasets}
