% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountryGroups}
\alias{CountryGroups}
\title{Membership in Country Groups}
\format{
\describe{
  \item{G8}{country belongs to G8}
  \item{G20}{country belongs to G20}
  \item{GGG}{country belongs to GGG (global governance group)}
}
}
\source{
http://en.wikipedia.org/wiki/G8,
https://www.g20.org/about_g20/g20_members ,
http://www.mfa.gov.sg/content/mfa/overseasmission/newyork/nyemb_statements/global_governance_group/2012/201209/press_20122809.html
}
\usage{
data("CountryGroups")
}
\description{
A data frame listing countries with logical variables indicating whether
each country belongs to the G8, the G20, or the GGG.
}
\details{
The country names are standardized to the CIA World Factbook names.
Each of the variables G8, G20, GGG is a logical variable. TRUE means the
country belongs to the corresponding group.
}
\examples{
data(CountryGroups)
str(CountryGroups)
}
\keyword{datasets}
