% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DirectRecoveryGroups}
\alias{DirectRecoveryGroups}
\title{Descriptions of the Direct Recovery Groups (DRGs) in the Medicare data.}
\format{
A data frame with 100 observations on the following 2 variables:
\describe{
  \item{drg}{a character string that looks like a number}
  \item{drgDefinition}{a text description of the DRG}
}
}
\source{
Extracted from the highly repetitive table provided by the Centers
for Medicare and Medicaid Services. See
http://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/Inpatient.html
}
\usage{
data("DirectRecoveryGroups")
}
\description{
Provides the medical meaning of the DRG codes.
}
\details{
This data table is related to MedicareCharges and MedicareProviders.
}
\examples{
head(DirectRecoveryGroups)
}
\keyword{datasets}
