% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HappinessIndex}
\alias{HappinessIndex}
\title{World Happiness Report Data}
\format{
A data frame with 156 observations on the following 13 variables:
\describe{
  \item{country}{name of the country}
  \item{region}{numerical code for a region of the world}
  \item{score}{The happiness index: a number from 0 to 10}
  \item{socialSupport}{A measure of having someone to count on}
  \item{freedom}{Perceived freedom to make life choices}
  \item{corruption}{Freedom from corruption}
  \item{donation}{}
  \item{generosity}{The generosity of people}
  \item{affectPos}{}
  \item{affectNeg}{}
  \item{happinessYesterday}{A presumably subjective measure of how happy the
  poll respondants were the day before the poll.}
  \item{gdpPerCapita}{In dollars}
  \item{lifeExpectancy}{The "healthy life expectancy" in years.}
}
}
\source{
The World Happiness Report is at
<http://unsdsn.org/wp-content/uploads/2014/02/WorldHappinessReport2013_online.pdf>
The original data used to create `HappinessIndex` are at
<http://unsdsn.org/wp-content/uploads/2013/09/Chapter-2_online-appendix_9-5-13_final.pdf>
}
\usage{
data("HappinessIndex")
}
\description{
The United Nations Sustainable Development Solutions Network
releases a report giving several indices of "happiness" measured for each of
more than 150 countries. These data are from the 2013 report
(<http://unsdsn.org/resources/publications/world-happiness-report-2013/>).
Underlying the data are the results of Gallop World Polls — the country data
is presumably the aggregate of data from individual people, although how the
aggregate is made is uncertain to this writer.
}
\details{
The healthy life expectancy (HALE) is a World Health Organization
measure of how long, on average, a person can be expected to live in good
health.

The data were scraped from a PDF file released by the United Nations. A few
country names were changed to match those in `CountryData` (e.g., Myanmar ->
Burma). Not all of the `HappinessIndex` countries are in `CountryData`, and vice
versa.
}
\examples{
data(HappinessIndex)
str(HappinessIndex)
}
\keyword{datasets}
