% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MigrationFlows}
\alias{MigrationFlows}
\title{Human Migration between Countries}
\format{
A data frame with 107184 observations on the following 8 variables:
\describe{
  \item{sex}{a factor with levels Female and Male}
  \item{destcode}{three letter country codes indicating the destination of the migrants}
  \item{origincode}{three letter country codes indicating where the migrants left}
  \item{Y2000}{Migration in year 2000. The value is the number of people
  (males and females separately) who migrated from the origin country to the destination country.}
  \item{Y1990}{Migration in year 1990}
  \item{Y1980}{Migration in year 1980}
  \item{Y1970}{Migration in year 1970}
  \item{Y1960}{Migration in year 1960}
}
}
\source{
The World Bank's Global Bilateral Migration Database:
\url{http://data.worldbank.org/data-catalog/global-bilateral-migration-databas}
}
\usage{
data("MigrationFlows")
}
\description{
The number of people who moved in from one country to another.
Data are provided for 1960, 1970, 1980, 1990, and 2000. There are separate
counts for males and females.
}
\examples{
data(MigrationFlows)
str(MigrationFlows)
}
\keyword{datasets}
