% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Minneapolis2017}
\alias{Minneapolis2017}
\title{Ballots in the 2017 Mayoral election in Minneapolis}
\format{
A data frame with 105,928 observations on the following 5 variables:
\describe{
  \item{precinct}{Precincts are sub-divisions within `wards`}
  \item{first}{The voter's first choice}
  \item{second}{The voter's second choice}
  \item{third}{The voter's third choice}
  \item{ward}{The city is divided spatially into districts or 'wards'. These
  are further subdivided into precincts}
}
}
\source{
Ballot data from the Minneapolis city government:
https://vote.minneapolismn.gov/results-data/election-results/2017/mayor-data/
}
\usage{
data("Minneapolis2017")
}
\description{
The choices marked on each (valid) ballot for the election,
which was run using a rank-choice, instant runoff system.
}
\details{
Ballot information for the 2017 Minneapolis Mayoral election, which
was run as a rank-choice election. In rank-choice, a voter can indicate
first, second, and third choices. If a voter's first choice is eliminated
(by being last in the count across voters), the second choice is promoted to
that voter's first choice, and similarly third -> second. Eliminations are
done successively until one candidate has a majority of the first-choice
votes. For additional detail, see
https://vote.minneapolismn.gov/ranked-choice-voting/details/
}
\examples{
data(Minneapolis2017)
str(Minneapolis2017)
}
\keyword{datasets}
