% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCHS}
\alias{NCHS}
\title{Health Statistics Data 1999-2004}
\format{
A data frame with 31126 observations on the following 31 variables:
\describe{
  \item{sex}{"male" or "female"}
  \item{age}{years}
  \item{pregnant}{yes" or "no"}
  \item{ethnicity}{Mexican American, Other Hispanic, Non-Hispanic White,
  Non-Hispanic Black, or Other/Multi}
  \item{death}{"alive", "cardiovascular death" or "other death"}
  \item{followup}{months of follow up (for mortality data)}
  \item{smoker}{"yes" or "no"}
  \item{diabetic}{"yes" or "no"}
  \item{height}{in meters}
  \item{weight}{in kilograms}
  \item{waist}{waist circumference (meters)}
  \item{wci}{the proposed body shape index}
  \item{bmi}{body mass index}
  \item{ptfat}{percent trunk fat}
  \item{tfat}{mass of trunk fat}
  \item{lfat}{limb fat}
  \item{llean}{limb lean tissue}
  \item{lbmi}{lean-tissue only BMI}
  \item{fbmi}{fat-only BMI}
  \item{bbmi}{bone BMI}
  \item{pfat}{percent fat}
  \item{bmd}{bone mineral density}
  \item{fmhm_other}{Framingham risk score}
  \item{hdl}{HDL cholesterol}
  \item{chol}{cholesterol (presumably, LDL)}
  \item{bps}{systolic blood pressure, mmHg}
  \item{bpd}{diastolic blood pressure, mmHg}
  \item{income}{ratio of family income to poverty threshold. 5 indicates a
  ratio greater than or equal to 5}
  \item{pop_weight}{population weight, for adding up over the US population}
  \item{psu}{primary sampling unit}
  \item{stratum}{sampling stratum}
}
}
\source{
These data were assembled from NCHS/NHANES 1999-2004 sources by
Dr. Nir Krakauer at City College of New York.
}
\usage{
data("NCHS")
}
\description{
These are survey data collected by the US National Center for
Health Statistics (NCHS). Each year, approximately 5,000 individuals are
interviewed in their homes. The data cover 1999-2004.

Note that the NHANES package provides data from 2009-2012, including more
health, life-style, and socio-economic variables
}
\examples{
data(NCHS)
str(NCHS)
}
\seealso{
The `NHANES` package available through CRAN.
}
\keyword{datasets}
