% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCI60cells}
\alias{NCI60cells}
\title{Cell Line descriptions in the NCI-60 dataset}
\format{
A data frame with 60 observations on the following 12 variables:
\describe{
  \item{cellLine}{Name from NCI-60 human tumor cell lines screen}
  \item{tissue}{The cancer tissue from which the cell line originated:
  Breast, CNS, Colon, Leukemia, Melanoma, Non-Small Cell Lung, Ovarian,
  Prostate, and Renal}
  \item{age}{of the cell line donor}
  \item{sex}{of the cell line donor}
  \item{ploidy}{the number of sets of chromosomes in the cell}
  \item{prior.treatment}{treatment of the donor}
  \item{epithelial}{is the cell of epithelial origin}
  \item{histology}{}
  \item{p53}{status of p53 protein believed to help regulate gene expression:
  mutated "MT", wild-type "WT", "?" }
  \item{mdr}{multidrug resistence (numeric)}
  \item{source}{information about where the cells were collected from}
  \item{doublingtime}{numeric}
}
}
\usage{
data("NCI60cells")
}
\description{
NCI60cells gives information about the cell-line tissue and the
person from whom the sample was collected.
}
\details{
These are the cell-line descriptions corresponding to the columns
in the `NCI60` expression dataset.
}
\examples{
data(NCI60cells)
str(NCI60cells)
}
\seealso{
`NCI60`
}
\keyword{datasets}
