% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrdwayBirds}
\alias{OrdwayBirds}
\title{Birds captured and released at Ordway, complete and uncleaned}
\format{
A data frame with 15,829 observations on the following 26 variables:
\describe{
  \item{bogus}{}
  \item{Timestamp}{indicates when the data were entered into an electronic
  record, not anything about the bird being described}
  \item{Year}{year of capture}
  \item{Day}{day of capture}
  \item{Month}{month of capture}
  \item{CaptureTime}{time of capture}
  \item{SpeciesName}{}
  \item{Sex}{}
  \item{Age}{}
  \item{BandNumber}{}
  \item{TrapID}{}
  \item{Weather}{}
  \item{BandingReport}{}
  \item{RecaptureYN}{}
  \item{RecaptureMonth}{}
  \item{RecaptureDay}{}
  \item{Condition}{}
  \item{Release}{}
  \item{Comments}{}
  \item{DataEntryPerson}{}
  \item{Weight}{}
  \item{WingChord}{}
  \item{Temperature}{}
  \item{RecaptureOriginal}{}
  \item{RecapturePrevious}{}
  \item{TailLength}{}
}
}
\source{
Jerald Dosch, Dept. of Biology, Macalester College: the manager of the Study Area.
}
\usage{
data("OrdwayBirds")
}
\description{
The historical record of birds captured and released at the
Katharine Ordway Natural History Study Area, a 278-acre preserve in Inver
Grove Heights, Minnesota, owned and managed by Macalester College.
}
\details{
There are many extraneous levels of variables such as species. Part
of the purpose of this data set is to teach about data cleaning.
}
\examples{
data(OrdwayBirds)
str(OrdwayBirds)
}
\seealso{
`OrdwaySpeciesNames`
}
\keyword{datasets}
