% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrdwaySpeciesNames}
\alias{OrdwaySpeciesNames}
\title{Corrected Species Names for the Ordway Birds}
\format{
A data frame with 265 observations on the following 2 variables:
\describe{
  \item{SpeciesName}{The original spelling, or misspelling, of a bird species.}
  \item{SpeciesNameCleaned}{Corrected spelling (or NA if the original was not identifiable.)}
}
}
\source{
Daniel Kaplan and students in a 2013 Data and Computing Fundamentals
class at Macalester College (Saint Paul, MN) read through original names
in `OrdwayBirds` and typed corrected spelling shown in `SpeciesNameCleaned`.
}
\usage{
data("OrdwaySpeciesNames")
}
\description{
This data frame lists all the species name that appear in
`OrdwayBirds`. In many cases, the species name was mis-spelled in the
original. As a result, many birds are listed as separate species even
though, in reality, they all belong to the same species. For each
potentially mis-spelled species name, this table gives a standardized name.
}
\examples{
data(OrdwaySpeciesNames)
str(OrdwaySpeciesNames)
}
\seealso{
`OrdwayBirds`
}
\keyword{datasets}
