% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegisteredVoters}
\alias{RegisteredVoters}
\title{A sample of the voter registration list for Wake County, North
Carolina in Fall 2010.}
\format{
A data frame with 10,000 observations on the following 19 variables:
\describe{
  \item{gender}{M = male, F = female, U = unknown}
  \item{party}{Political affiliation: DEM = Democratic, LIB = Libertarian, REP = Republican, UNA = Unaffiliated}
  \item{Age}{age in years}
  \item{race}{race: A = Asian, B = black or African American, I = American Indian or Alaska Native,
  M = More than one race, O = Other, U = Undesignated, W = White}
  \item{voter_reg_num}{Unique serial number for each voter}
  \item{last_name}{Voter's last name}
  \item{first_name}{Voter's first name}
  \item{midl_name}{Voter's middle name}
  \item{name_sufx}{Suffix, if any, on the voter's name}
  \item{mail_city}{Component of voter's mailing address}
  \item{mail_state_cd}{Component of voter's mailing address}
  \item{mail_street}{Component of voter's mailing address}
  \item{mail_zip_code}{Component of voter's mailing address}
  \item{registr_dt}{Component of voter's mailing address}
  \item{res_city}{Component of voter's mailing address}
  \item{res_state_cd}{Component of voter's mailing address}
  \item{res_unit_num}{Component of voter's mailing address}
  \item{res_zip_code}{Component of voter's mailing address}
  \item{voter_status}{Status of the voter}
}
}
\source{
The data file is publicly available from the Wake County (NC) Board
of Elections. http://msweb03.co.wake.nc.us/bordelec/Waves/WavesDownload.asp.
It's in a Windows self-extracting format. Prof. Shilad Sen of Macalester
College (Saint Paul, MN) reformatted the data to accompany the DataComputing
book.

The full data set, containing 583,092 records, is available at
http://www.mosaic-web.org/go/Repository/DannyKaplan/WakeCountyBOE.Rdata
You can access it by downloading the file, then loading it into your R
session. The data frame is `WakeCountyBOE`.
}
\usage{
data("RegisteredVoters")
}
\description{
These are data from the Wake County Board of Elections. As so
often happens, the web sites with the documentation and data are no longer
current.
}
\details{
For additional fields, see the original documentation at
http://msweb03.co.wake.nc.us/bordelec/Waves/boedatadescription.pdf
}
\examples{
data(RegisteredVoters)
str(RegisteredVoters)
}
\keyword{datasets}
