% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldCities}
\alias{WorldCities}
\title{Cities and their populations}
\format{
A data frame with 23018 observations on the following 10 variables:
\describe{
  \item{code}{The ISO (?) city code}
  \item{name}{Name of the city}
  \item{latitude}{location in degrees}
  \item{longitude}{location in degrees}
  \item{country}{Two letter country code}
  \item{countryRegion}{A numerical region}
  \item{population}{}
  \item{regionCode}{ISO (?) Code}
  \item{region}{Name of the region}
  \item{date}{Date estimate made}
}
}
\usage{
data("WorldCities")
}
\description{
A list of world cities accomanied by location information.
}
\examples{
data(WorldCities)
str(WorldCities)
}
\keyword{datasets}
