% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZipDemography}
\alias{ZipDemography}
\title{Demographic information for most US ZIP Codes (Postal Codes)}
\format{
A data frame with 42741 observations on the following 44 variables:
\describe{
  \item{Totalpopulation}{}
  \item{Male}{}
  \item{Female}{}
  \item{MedianAge}{}
  \item{Under5years}{}
  \item{X18yearsandover}{}
  \item{X65yearsandover}{}
  \item{Onerace}{}
  \item{White}{}
  \item{BlackorAfricanAmerican}{}
  \item{AmericanIndianandAlaskaNative}{}
  \item{Asian}{}
  \item{NativeHawaiianandOtherPacificIslander}{}
  \item{Someotherrace}{}
  \item{Twoormoreraces}{}
  \item{HispanicorLatinoofanyrace}{}
  \item{AverageHouseholdSize}{}
  \item{Averagefamilysize}{}
  \item{Totalhousingunits}{}
  \item{Occupiedhousingunits}{}
  \item{Owneroccupiedhousingunits}{}
  \item{Renteroccupiedhousingunits}{}
  \item{Vacanthousingunits}{}
  \item{Population25yearsandover}{}
  \item{Highschoolgraduateorhigher}{}
  \item{Bachelorsdegreeorhigher}{}
  \item{Civilianveterans}{}
  \item{Disabilitystatuspopulation21to64years}{}
  \item{Foreignborn}{}
  \item{Nowmarriedpopulation15yearsandover}{}
  \item{SpeakalanguageotherthanEnglishathome5yearsandover}{}
  \item{Inlaborforcepopulation16yearsandover}{}
  \item{Meantraveltimetoworkinminutespopulation16yearsandolder}{}
  \item{Medianhouseholdincomedollars}{}
  \item{Medianfamilyincomedollars}{}
  \item{Percapitaincomedollars}{}
  \item{Familiesbelowpovertylevel}{}
  \item{Individualsbelowpovertylevel}{}
  \item{Singlefamilyowneroccupiedhomes}{}
  \item{Medianvaluedollars}{}
  \item{Medianofselectedmonthlyownercosts}{}
  \item{WithaMortgage}{}
  \item{Notmortgaged}{}
  \item{ZIP}{}
}
}
\usage{
data("ZipDemography")
}
\description{
Data on the population in each of ZIP code region.
}
\examples{
data(ZipDemography)
str(ZipDemography)
}
\keyword{datasets}
