% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZipGeography}
\alias{ZipGeography}
\title{Geographic information by US Zip Codes (Postal Codes)}
\format{
A data frame with 42741 observations on the following 13 variables:
\describe{
  \item{State}{}
  \item{Population}{the population of the ZIP area}
  \item{County}{the name of the county to which the ZIP area belongs}
  \item{HousingUnits}{}
  \item{LandArea}{}
  \item{WaterArea}{}
  \item{CityName}{}
  \item{AreaCode}{}
  \item{Timezone}{}
  \item{Latitude}{}
  \item{Longitude}{}
  \item{PostOfficeName}{the name of the post office serving the ZIP}
  \item{ZIP}{character vector for the ZIP code area}
}
}
\usage{
data("ZipGeography")
}
\description{
Information about the location, area, and housing units in each ZIP code area.
}
\examples{
data(ZipGeography)
str(ZipGeography)
}
\keyword{datasets}
